examples/python-stack/requirements.txt
aws-cdk-lib~=2.110.1
aws-cdk.aws-lambda-go-alpha~=2.110.1a0
constructs~=10.0.5
datadog_cdk_constructs_v2~=1.9.0
